Fix "similar-code" issue in examples/clients/clients.java
