+++ examples/clients/RetrieveStats.java
@@ -7,37 +7,13 @@ import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 
-//////////////////////////////////////////////////////////////////
-// Retrieve email statistics by client type.
-// GET /clients/stats
-
-
-public class Example {
-  public static void main(String[] args) throws IOException {
-    try {
-      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
-      Request request = new Request();
-      request.setMethod(Method.GET);
-      request.setEndpoint("clients/stats");
-      request.addQueryParam("aggregated_by", "day");
-      request.addQueryParam("start_date", "2016-01-01");
-      request.addQueryParam("end_date", "2016-04-01");
-      Response response = sg.api(request);
-      System.out.println(response.getStatusCode());
-      System.out.println(response.getBody());
-      System.out.println(response.getHeaders());
-    } catch (IOException ex) {
-      throw ex;
-    }
-  }
-}
 
 //////////////////////////////////////////////////////////////////
 // Retrieve stats by a specific client type.
 // GET /clients/{client_type}/stats
 
 
-public class Example {
+public class RetrieveStats {
   public static void main(String[] args) throws IOException {
     try {
       SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
